Move from `utils` to `bulk`
