+++ corehq/apps/hqcase/utils.py
@@ -1,6 +1,5 @@
 import datetime
 import uuid
-from contextlib import contextmanager
 from xml.etree import cElementTree as ElementTree
 
 from django.template.loader import render_to_string
@@ -109,98 +108,6 @@ def submit_case_blocks(
         **submission_extras
     )
     return result.xform, result.cases
-
-
-def coro_as_context(func):
-    """
-    Decorator to transform a coroutine function into a context manager.
-
-    The context manager primes the coroutine using ``next()`` when
-    entering the context, and calls ``.close()`` when exiting the context.
-
-    Usage::
-
-        @coro_as_context
-        def my_coro():
-            try:
-                while True:
-                    data = yield
-                    ... # Do something with data
-            except GeneratorExit:
-                ... # Cleanup
-
-        with my_coro() as coro:
-            coro.send(data)
-
-    """
-    @contextmanager
-    def wrapper(*args, **kwargs):
-        coro = func(*args, **kwargs)
-        next(coro)  # Prime the coroutine
-        try:
-            yield coro
-        finally:
-            coro.close()
-    return wrapper
-
-
-def submit_case_block_coro(
-    *args,
-    chunk_size=CASEBLOCK_CHUNKSIZE,
-    **kwargs,
-):
-    """
-    Accepts case blocks and submits them in chunks of chunk_size.
-    Returns a list of form IDs.
-
-    Use undecorated as a coroutine for access to the return value, or
-    use `submit_case_block_context()` for simplicity.
-
-    Context manager usage::
-
-        with submit_case_block_context(domain, device_id=__name__) as submit:
-            for case in iter_all_the_cases:
-                case_block = get_case_updates(case)
-                submit.send(case_block)
-
-    Coroutine usage::
-
-        submit = submit_case_block_coro(domain, device_id=__name__)
-        next(submit)  # Prime the coroutine
-        for case in iter_all_the_cases:
-            case_block = get_case_updates(case)
-            submit.send(case_block)
-        try:
-            submit.send(None)  # Trigger exit
-        except StopIteration as exc:
-            form_ids = exc.value
-
-    """
-    form_ids = []
-    case_blocks = []
-    try:
-        while True:
-            case_block = yield
-            if case_block is None:  # Send `None` to exit
-                raise GeneratorExit
-            case_blocks.append(
-                case_block.as_text()
-                if isinstance(case_block, CaseBlock)
-                else case_block
-            )
-            if len(case_blocks) >= chunk_size:
-                chunk = case_blocks[:chunk_size]
-                case_blocks = case_blocks[chunk_size:]
-                xform, __ = submit_case_blocks(chunk, *args, **kwargs)
-                form_ids.append(xform.form_id)
-    except GeneratorExit:
-        if case_blocks:
-            xform, __ = submit_case_blocks(case_blocks, *args, **kwargs)
-            form_ids.append(xform.form_id)
-    return form_ids
-
-
-submit_case_block_context = coro_as_context(submit_case_block_coro)
 
 
 def get_case_by_identifier(domain, identifier):
